Fixing most of the css
Still a few halfassed things with the css that I can't figure out how to fix.

Most things work for every preview screen in the inspect tab with a few small exceptions.

The Navbar is fixes so that you can actually use it with the phone and ipad (m.m.)